requirements.txt
fastapi
openai-whisper==20231117
torch==2.1.2
transformers[torch]
librosa==0.10.2.post1
numpy
uvicorn
openai==1.35.14
python-dotenv
scikit-learn==1.3.0
pandas
boto3==1.35.10